Update GUI feature
