--- data/jag.txt
@@ -1,2 +1,0 @@
-[T][X] read magazine
-[T][X] read magazine
