+++ src/main/java/jag/MainWindow.java
@@ -5,8 +5,8 @@ import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
-import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
+import javafx.application.Platform;
 
 /**
  * Controller for the main GUI
@@ -28,7 +28,7 @@ public class MainWindow {
 
     @FXML
     public void initialize() {
-        String greetings = Ui.getGreetings();
+        String greetings = UiGUI.getGreetings();
         scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
         dialogContainer.getChildren().add(DialogBox.getJagDialog(greetings, jagImage));
     }
@@ -54,5 +54,10 @@ public class MainWindow {
                 DialogBox.getJagDialog(response, jagImage)
         );
         userInput.clear();
+
+        // Stop the JavaFX application from running after bye command
+        if (response.contains("Bye")) {
+            Platform.exit();
+        }
     }
 }
